@@ -3,36 +3,28 @@
  * @author Wei-Ming Wu
  *
  *
- * Copyright 2014 Wei-Ming Wu
+ *         Copyright 2014 Wei-Ming Wu
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
+ *         Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ *         except in compliance with the License. You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *         http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
+ *         Unless required by applicable law or agreed to in writing, software distributed under the
+ *         License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ *         either express or implied. See the License for the specific language governing
+ *         permissions and limitations under the License.
  *
  */
 package com.github.wnameless.smartcard;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertTrue;
-import static org.junit.Assert.fail;
-
+import static org.junit.jupiter.api.Assertions.*;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Modifier;
 import java.nio.ByteBuffer;
-
 import javax.smartcardio.CommandAPDU;
-
-import org.junit.Before;
-import org.junit.Test;
-
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
 import com.github.wnameless.smartcard.APDU.APDUBuilder;
 import com.google.common.primitives.Bytes;
 import com.google.common.testing.NullPointerTester;
@@ -41,7 +33,7 @@ public class APDUTest {
 
   APDUBuilder builder;
 
-  @Before
+  @BeforeEach
   public void setUp() throws Exception {
     builder = APDU.builder();
   }
@@ -74,60 +66,59 @@ public class APDUTest {
 
   @Test
   public void testDefaultAPDU() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00 }), builder.build());
+    assertEquals(new CommandAPDU(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00}),
+        builder.build());
   }
 
   @Test
   public void testSetCLA() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x01, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00 }), builder.setCLA((byte) 0x01).build());
+    assertEquals(new CommandAPDU(new byte[] {(byte) 0x01, (byte) 0x00, (byte) 0x00, (byte) 0x00}),
+        builder.setCLA((byte) 0x01).build());
   }
 
   @Test
   public void testSetINS() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, INS.APPEND_RECORD,
-        (byte) 0x00, (byte) 0x00 }), builder.setINS(INS.APPEND_RECORD).build());
+    assertEquals(
+        new CommandAPDU(new byte[] {(byte) 0x00, INS.APPEND_RECORD, (byte) 0x00, (byte) 0x00}),
+        builder.setINS(INS.APPEND_RECORD).build());
   }
 
   @Test
   public void testSetP1() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x01, (byte) 0x00 }), builder.setP1((byte) 0x01).build());
+    assertEquals(new CommandAPDU(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x00}),
+        builder.setP1((byte) 0x01).build());
   }
 
   @Test
   public void testSetP2() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x02 }), builder.setP2((byte) 0x02).build());
+    assertEquals(new CommandAPDU(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x02}),
+        builder.setP2((byte) 0x02).build());
   }
 
   @Test
   public void testSetData() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x11 }),
+    assertEquals(new CommandAPDU(
+        new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x11}),
         builder.setData((byte) 0x11).build());
+    assertEquals(new CommandAPDU(
+        Bytes.concat(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0xFF},
+            new byte[255])),
+        builder.setData(new byte[255]).build());
     assertEquals(
-        new CommandAPDU(Bytes.concat(new byte[] { (byte) 0x00, (byte) 0x00,
-            (byte) 0x00, (byte) 0x00, (byte) 0xFF }, new byte[255])), builder
-            .setData(new byte[255]).build());
-    assertEquals(
-        new CommandAPDU(Bytes.concat(new byte[] { (byte) 0x00, (byte) 0x00,
-            (byte) 0x00, (byte) 0x00, (byte) 0x00 }, ByteBuffer.allocate(2)
-            .putShort((short) 256).array(), new byte[256])),
+        new CommandAPDU(Bytes.concat(
+            new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00},
+            ByteBuffer.allocate(2).putShort((short) 256).array(), new byte[256])),
         builder.setData(new byte[256]).build());
   }
 
   @Test
   public void testSetDataWithHexString1() {
     assertEquals(
-        APDU.builder()
-            .setINS(INS.SELECT_FILE)
-            .setP1((byte) 0x04)
-            .setData((byte) 0xD1, (byte) 0x58, (byte) 0x00, (byte) 0x00,
-                (byte) 0x01, (byte) 0x00, (byte) 0x00, (byte) 0x00,
-                (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00,
-                (byte) 0x00, (byte) 0x00, (byte) 0x11, (byte) 0x00).build(),
+        APDU.builder().setINS(INS.SELECT_FILE).setP1((byte) 0x04)
+            .setData((byte) 0xD1, (byte) 0x58, (byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x00,
+                (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00,
+                (byte) 0x00, (byte) 0x00, (byte) 0x11, (byte) 0x00)
+            .build(),
         APDU.builder().setINS(INS.SELECT_FILE).setP1((byte) 0x04)
             .setData("D1580000010000000000000000001100").build());
   }
@@ -135,36 +126,35 @@ public class APDUTest {
   @Test
   public void testSetDataWithHexString2() {
     assertEquals(
-        APDU.builder()
-            .setINS(INS.SELECT_FILE)
-            .setP1((byte) 0x04)
-            .setData((byte) 0x00, (byte) 0x58, (byte) 0x00, (byte) 0x00,
-                (byte) 0x01, (byte) 0x00, (byte) 0x00, (byte) 0x00,
-                (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00,
-                (byte) 0x00, (byte) 0x00, (byte) 0x11, (byte) 0x00).build(),
+        APDU.builder().setINS(INS.SELECT_FILE).setP1((byte) 0x04)
+            .setData((byte) 0x00, (byte) 0x58, (byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x00,
+                (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00,
+                (byte) 0x00, (byte) 0x00, (byte) 0x11, (byte) 0x00)
+            .build(),
         APDU.builder().setINS(INS.SELECT_FILE).setP1((byte) 0x04)
             .setData("00580000010000000000000000001100").build());
   }
 
   @Test
   public void testSetDataWithInsufficiencyHexString() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x10 }),
+    assertEquals(new CommandAPDU(
+        new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x01, (byte) 0x10}),
         builder.setData("1").build());
   }
 
-  @Test(expected = IllegalArgumentException.class)
+  @Test
   public void testSetDataWithInvalidHexString() {
-    builder.setData("qerb11").build();
+    assertThrows(IllegalArgumentException.class, () -> {
+      builder.setData("qerb11").build();
+    });
   }
 
   @Test
   public void testSetLongData() {
     byte[] longData = new byte[512];
-    assertEquals(
-        new CommandAPDU(Bytes.concat(new byte[] { (byte) 0x00, (byte) 0x00,
-            (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x02, (byte) 0x00 },
-            longData)), builder.setData(longData).build());
+    assertEquals(new CommandAPDU(Bytes.concat(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00,
+        (byte) 0x00, (byte) 0x00, (byte) 0x02, (byte) 0x00}, longData)),
+        builder.setData(longData).build());
   }
 
   @Test
@@ -185,24 +175,24 @@ public class APDUTest {
 
   @Test
   public void testClearData() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00 }), builder.setData((byte) 0x11).clearData()
-        .build());
+    assertEquals(new CommandAPDU(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00}),
+        builder.setData((byte) 0x11).clearData().build());
   }
 
   @Test
   public void testSetLe() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00, (byte) 0x10 }), builder.setLe(16).build());
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x02, (byte) 0x00 }),
-        builder.setLe(512).build());
+    assertEquals(
+        new CommandAPDU(
+            new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x10}),
+        builder.setLe(16).build());
+    assertEquals(new CommandAPDU(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00,
+        (byte) 0x00, (byte) 0x02, (byte) 0x00}), builder.setLe(512).build());
   }
 
   @Test
   public void testClearLe() {
-    assertEquals(new CommandAPDU(new byte[] { (byte) 0x00, (byte) 0x00,
-        (byte) 0x00, (byte) 0x00 }), builder.setLe(512).clearLe().build());
+    assertEquals(new CommandAPDU(new byte[] {(byte) 0x00, (byte) 0x00, (byte) 0x00, (byte) 0x00}),
+        builder.setLe(512).clearLe().build());
   }
 
   @Test
