@@ -3,29 +3,25 @@
  * @author Wei-Ming Wu
  *
  *
- * Copyright 2014 Wei-Ming Wu
+ *         Copyright 2014 Wei-Ming Wu
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
+ *         Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ *         except in compliance with the License. You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *         http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
+ *         Unless required by applicable law or agreed to in writing, software distributed under the
+ *         License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ *         either express or implied. See the License for the specific language governing
+ *         permissions and limitations under the License.
  *
  */
 package com.github.wnameless.smartcard;
 
-import static org.junit.Assert.assertTrue;
-
+import static org.junit.jupiter.api.Assertions.assertTrue;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Modifier;
-
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 public class INSTest {
 
