@@ -3,41 +3,32 @@
  * @author Wei-Ming Wu
  *
  *
- * Copyright 2014 Wei-Ming Wu
+ *         Copyright 2014 Wei-Ming Wu
  *
- * Licensed under the Apache License, Version 2.0 (the "License"); you may not
- * use this file except in compliance with the License. You may obtain a copy of
- * the License at
+ *         Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file
+ *         except in compliance with the License. You may obtain a copy of the License at
  *
- * http://www.apache.org/licenses/LICENSE-2.0
+ *         http://www.apache.org/licenses/LICENSE-2.0
  *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations under
- * the License.
+ *         Unless required by applicable law or agreed to in writing, software distributed under the
+ *         License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+ *         either express or implied. See the License for the specific language governing
+ *         permissions and limitations under the License.
  *
  */
 package com.github.wnameless.smartcard;
 
 import static com.google.common.base.Preconditions.checkArgument;
-
 import java.nio.ByteBuffer;
-
 import javax.smartcardio.CommandAPDU;
-
-import net.sf.rubycollect4j.util.ByteUtil;
-
-import com.github.wnameless.nullproof.annotation.RejectNull;
 import com.google.common.primitives.Bytes;
+import net.sf.rubycollect4j.util.ByteUtils;
 
 /**
  * 
- * {@link APDU} is a helper class designed to make the creation of a CommandAPDU
- * easier.
+ * {@link APDU} is a helper class designed to make the creation of a CommandAPDU easier.
  *
  */
-@RejectNull
 public final class APDU {
 
   private APDU() {}
@@ -53,11 +44,9 @@ public final class APDU {
 
   /**
    * 
-   * {@link APDUBuilder} is designed to hold input data from user before a
-   * CommandAPDU is created.
+   * {@link APDUBuilder} is designed to hold input data from user before a CommandAPDU is created.
    *
    */
-  @RejectNull
   public static class APDUBuilder {
 
     private final byte[] apdu = new byte[4];
@@ -68,8 +57,7 @@ public final class APDU {
     /**
      * Sets the CLA byte, default is 0x00.
      * 
-     * @param claByte
-     *          a CLA byte
+     * @param claByte a CLA byte
      * @return this {@link APDUBuilder}
      */
     public APDUBuilder setCLA(byte claByte) {
@@ -80,8 +68,7 @@ public final class APDU {
     /**
      * Sets the INS byte, default is 0x00.
      * 
-     * @param insByte
-     *          a INS byte
+     * @param insByte a INS byte
      * @return this {@link APDUBuilder}
      */
     public APDUBuilder setINS(byte insByte) {
@@ -92,8 +79,7 @@ public final class APDU {
     /**
      * Sets the P1 byte, default is 0x00.
      * 
-     * @param p1Byte
-     *          a P1 byte
+     * @param p1Byte a P1 byte
      * @return this {@link APDUBuilder}
      */
     public APDUBuilder setP1(byte p1Byte) {
@@ -104,8 +90,7 @@ public final class APDU {
     /**
      * Sets the P2 byte, default is 0x00.
      * 
-     * @param p2Byte
-     *          a byte
+     * @param p2Byte a byte
      * @return this {@link APDUBuilder}
      */
     public APDUBuilder setP2(byte p2Byte) {
@@ -121,7 +106,7 @@ public final class APDU {
     private byte[] lengthOfData(int length) {
       byte[] lenBytes;
       if (length < 256) {
-        lenBytes = new byte[] { (byte) length };
+        lenBytes = new byte[] {(byte) length};
       } else {
         ByteBuffer buffer = ByteBuffer.allocate(3);
         buffer.position(1);
@@ -140,8 +125,7 @@ public final class APDU {
      * Sets the DATA bytes, default is empty.<br>
      * If the DATA is set, a Lc bytes is set automatically as well.
      * 
-     * @param dataBytes
-     *          the DATA bytes
+     * @param dataBytes the DATA bytes
      * @return this {@link APDUBuilder}
      */
     public APDUBuilder setData(byte... dataBytes) {
@@ -156,14 +140,12 @@ public final class APDU {
      * Sets the DATA bytes by given HEX string, default is empty.<br>
      * If the DATA is set, a Lc bytes are set automatically as well.
      * 
-     * @param hexString
-     *          a hex string
+     * @param hexString a hex string
      * @return this {@link APDUBuilder}
-     * @throws IllegalArgumentException
-     *           if hexadecimal string is invalid
+     * @throws IllegalArgumentException if hexadecimal string is invalid
      */
     public APDUBuilder setData(String hexString) {
-      byte[] bytes = ByteUtil.fromHexString(hexString);
+      byte[] bytes = ByteUtils.fromHexString(hexString);
       if (bytes.length == 0) {
         lc = null;
         data = null;
@@ -188,13 +170,11 @@ public final class APDU {
     /**
      * Sets the Le bytes, default is empty.
      * 
-     * @param leLength
-     *          length of Le
+     * @param leLength length of Le
      * @return this {@link APDUBuilder}
      */
     public APDUBuilder setLe(int leLength) {
-      checkArgument(leLength >= 1 && leLength <= 65535,
-          "Le is between 1..65535");
+      checkArgument(leLength >= 1 && leLength <= 65535, "Le is between 1..65535");
       le = lengthOfData(leLength);
       return this;
     }
@@ -216,10 +196,8 @@ public final class APDU {
      */
     public CommandAPDU build() {
       byte[] finalApdu = apdu;
-      if (lc != null)
-        finalApdu = Bytes.concat(apdu, lc, data);
-      if (le != null)
-        finalApdu = Bytes.concat(finalApdu, le);
+      if (lc != null) finalApdu = Bytes.concat(apdu, lc, data);
+      if (le != null) finalApdu = Bytes.concat(finalApdu, le);
       return new CommandAPDU(finalApdu);
     }
 
